Refactor: 로그인 로직 분리 및 보안 강화
- 로그인 로직을 useAuth 커스텀 훅으로 분리
- 환경 변수 기반 동적 URL 처리 개선
- 민감 정보 로깅 제거 및 개발 환경만 로그 출력
